Feat: 사이드바 help 누를 시 모달 구현
